-add 2 new prefetch for house interiors

+++ zeditor/src/zeditor/core/prefetch/patch/CastleLow.java
@@ -3,11 +3,11 @@ package zeditor.core.prefetch.patch;
 
 public class CastleLow extends AbstractPatch12 {
 
-	byte[] conv_value = 
+	int[] conv_value = 
 	{ 42, 46, 44, 45, 39, 43, -1, 53, 37, -1, 41, 52, 38, 51, 50, 42 };
 
-	byte[] value =
-	getReverseTab(conv_value, 37);
+	int[] value =
+		getReverseTab(conv_value, 37);
 
 	public CastleLow() {
 		super(true);
